refactor(user): reorganize user profile interfaces and update controller dependencies
