+++ src/presentation/routes/users/profile.route.ts
@@ -2,6 +2,7 @@ import { NextFunction, Request, Response } from "express";
 import {
   authMiddleware,
   getUserController,
+  s3Controller,
   updateProfileController,
 } from "../../../infrastructure/di/resolver";
 import { BaseRoute } from "../base.route";
@@ -35,7 +36,7 @@ export class UserRoutes extends BaseRoute {
       (req: Request, res: Response, next: NextFunction) =>
         authMiddleware.authenticate(req, res, next),
       (req: Request, res: Response, next: NextFunction) => {
-        updateProfileController.handle(req, res, next);
+        s3Controller.getUploadURL(req, res, next);
       }
     );
   }
